Fix for package and subpackage doc tags
Changed the format of the package and subpackage doc tags so that the
package now contains the name of the system (underscored plural of the
controler name) and the subpackage corresponds if it's a controller,
model or view of that system.

Since the view templates generate files for both admin and non admin
views a generic subpackage extension is used. They would need to be
changed manually.

The format of the doc tags is now:

- Models: system_name.model
- Controllers: system_name.controller
- Index view: system_name.views.index
- View view: system_name.views.view
- Add and edit view: system_name.views.form

Also changed the doc tags of this template to match the new format.