New AddThis elements.
Adding new configurable AddThis elements, one for a basic toolbar and
the other for posts. Both need a pubId defined, plus the one for posts
needs the configuration key posts set to true. The other common
settings are:

 - title to use, could be name of the site or twitter name.
 - boolean to specify using the style with 32x32 icons.

The post element accepts the slug to use and the post title to use.

Both elements use full URLs and include a small Javascript
configuration to supress clackback URL fragments.