New Albums system
Added new image albums system. It creates a folder for the images and
thumbnails. Images can be uploaded via FTP or one by one via a form.
Images and thumbnail names must match, they are diferentiated by their
location: thumbnails reside in the 'thumbnails' folder of the album.
Individual images can be deleted in the admin view view which displays
the thumbnails.

It uses the Slimbox2 script to show images ina lightbox. It must be
enabled by loading the plugin (located in the cake plugins folder)
in the project bootstrap.