Fix HTML5 markup for better use of header tags
Fixed frontend views HTML5 markup for better use of header tags.
Mainly moved info that was in divs (usually marked with the
content-info class) and moved it into the header section of the
articles. In the case of posts also moved the author to the header.
In the case of events moved the event location and times to the
header. In general moved the creation (publication) times to the
header.