Add configuration options for min. and max. year for date selects
Add functionality to set minimum and maximum years in date selects
of Events.In admin_add view both min. and max. year are set, in
admin_edit only max, year.

It determines if it should use the min. and max. years if set or use
the date ranges from the current year if they're not set.