Add comment about handling view actions with hasMany relationships
Added a comment about handling view actions with hasMany relationships.
It has an example from the Group hasMany User relationship. Besides
setting that in the admin_view action, the admin_view view has to be
modified to handle the pagination.

TODO: try to figure a way to automatically handle this, see if the
hasMany relationship is passed to the templates.