First version of Contact Form Emails system
Added first version of Contact Form Emails system, to be able to send
contact form messages to multiple emails without having to modify the
contact form each time a change is needed, since emails are stored in
the database. Emails have a boolean field to indicate if the email is
active or not.

System uses the custom find type "active" that returns a simple array
of emails for the contact form.