Add custom find type 'latest'
Add custom find type 'latest' and deleted the getLatest method.

Latest posts are retrieved via find('latest') and the number to
retrieve is specified in a configuration key. After retrieving results
in the _findLatest method, they're cached. The cached is not checked
here.

This means the cache has to be checked elsewhere like the controller
methods. In the case the latest possts cache is checked in the
latestPosts and index (RSS check) methods first, and if the cache is
expired or invalid a call to find('latest') is made.