Add functionality to add page slugs to link fields
Added functionality to add page slugs to link fields. The admin add
and edit pages list all available pages and has a button to add the
appropiate link.

The Menus controller needed to add a $uses variable to add the Page
model. The pages are set in a variable in the admin_add and admin_edit
methods. Each button to add a page slug has a menu-page-slug class to
capture click events and the is set to the page slug.

The admin Javascript has a new handler that captures events for
elements with the menu-page-slug class, gets the id of the clicked
button and set the link field to that value (adding the '/pages/view/'
text beforehand).