Modify Events and Posts system to not use requestAction() in elements
Modified the Events and Posts system to not use requestAction() in
the elements for the latest posts and upcoming events.

Eliminated the getLatest (in Posts) and getUpcoming (in Events)
methods, and eliminated the requestAction() calls in the corresponding
elements. Then added the Post and Event model to the Pages controller
and used the appropiate custom find methods, find('latest') in posts
and find('upcoming') in events, to get the data and set the appropiate
view variables.

Had to pass the view variables in the Pages index view to the elements.