Add functionality to save Page Sections with Pages
Added functionality to save Page Sections with Pages. In the
controller use the saveAssociated (with atomic false) call. In the
admin_edit view use a foreach loop to put the form inputs for the
page sections. In the admin_view view use a foreach loop to display
the page sections content.