Add Brita component to clean page sections
Added the Brita component to clean the page section when adding or
editing a page.

The page sections are purified, using the request data array, in a loop
prior to saving.

When the page is saved via the saveAssociated call, the result returned
is an array instead of a boolean, so added a check to check first if
the page ssaved correctly and then check if the page sections also
saved correctly, and display the appropiate message.

Additionally added the 'validate' => true setting in the saveAssociated
call in the admin_add method. This ensures that each page section is
validated once at a time upon saving, instead of all at the same time
before the page_id is set.