Simplify pagination and use new styles for paginator links
Simplified the pagination in all views. Deleted the first() and
last() Paginator Helper calls and added those options to numbers()
Paginator Helper method.

Also added the classes 'prev disabled' and 'next disabled' to the
disabled options in the next and prev methods of the Paginator Helper.

Modified the project CSS to take into account this changes and reflect
the styles of the Bootstrap from Twitter pagination components. In
particular the link tags are now the ones with the large hit area to
help in navigation.