Add WEBSITE_BASE_URL definition for use in shells
Changed the redefinition of FULL_BASE_URL in the project configuration
file to WEBSITE_BASE_URL to be used in shells and other places where
FULL_BASE_URL might not be properly set, and to not affect it.

This fixes the issue with login in the test environment (localhost)
that made the redirect fail.