Add Google Custom Search
Improved the search capabilities by using Google Custom Search.

Created a controller with no model so as to have a view where to
display the results. The index view includes an element with the
Google code to display the custom results which has a configuration
setting for the code.

Created two new elements one that displays the search box, which
is used in the default layout, and a custom search results element
that contains the Google Javascript code.

The system is configured via three settings in the project
configuration file:
- action: the action to display the results, defaults to the Google
  site. If empty it uses the index action of the SearchResults
  controller.
- sitesearch: in case of using a simple search, the site to search.
- key: the Google Custom Search code (site should be Google verified).