Add support for pages files folders
Added support for pages files folders. The admin_deleteFile has a new
parameter to specify the base location of the file, that defaults to
images. If the parameters doesn't match 'images' or 'files', no file
is deleted. This parameter is sent from the views as a simple string
to overcome the problem of URL encoded locations which might not work.
The model deleteFile method is called.

Also modified the admin_view and admin_edit functions to list the
already uploaded page files and added the ability to also upload one
file in the admin_edit view/method.