Add first version of FilesList datasource
Added first version of FilesList datasource. It obtains a list of
files from the specified directory and can have a configured base
path. The main advantage is that it allows for paginated results,
instead of a big array. The returned results are in an array
named after the model alias.

It contains two functions to paginate and sort the results. The
paginated results are done by calculating and offset and slicing the
array.

It'll be used mainly for the Album system which display all images
in a directory.