Modify Albums system to use AlbumImage model
Modified the Albums system to use AlbumImage model, whcih allows
for paginated results. Added the Album hasMany AlbumImage
relationship in the model, which helps in finding the model in the
controller.

In the controller we setup the paginate configuration in the view
method to setup the appropiate directory and set the limit to 60 for
pagination, since each row by default shows 6 thumbnails, this was
we can get 10 row. We then call the paginate function on the Album
Image model.

In the view view modified it to use the correct view variable,
$albumImages, and set the pagination controls in a separate row.