Add support for mobile sites
Added support for mobile sites using View Themes and Bootstrap's
responsive features.

Added support for mobile device detection and themes in the
beforeFilter callback. First we check if we have a session
variable set for the theme to use. It it isn't set we check to see
if the request is from a mobile device/browser and set the theme
variable accordingly.

We then check the theme session variable and set the Vire Theme
accordingly.

We also set the Vary response header since content my differ depending
on user agent.

Added first version of the Change Themes controller. It uses a simple
method to switch the theme session variable.

Updated the default layout to prepare for responsive mobile version.
Added the meta tag for viewport scale necessary for Ipad and
Iphones. Added the bootstrap collapse plugin, necessary for the menu.
Added the bootstrap-transition Javascript plugin to smooth out the
collapsible responsive menu.

Also added a button at the bottom to switch themes.

For the layout use the responsive project CSS version and use the
extra markup to make the menu collapsable. Also simplified the site
header.