Update login and logout redirection
Updated the login and logout redirection to use more modern CakePHP
conventions. In the AppController deleted the code that nullified
the login redirection in sessions. Added the Auth loginAction,
loginRedirect and logoutRedirect calls.

In the Users controller simplified the admin login and logout
methods by using the Auth redirect call and the Auth logout in the
redirect call respectively.