Modify pagination to use tags and styles for Bootstrap
Modified the numbers, prev and next pagination elements so that
they can be used with the default Bootstrap pagination styles. For
all elements the default tag was changed to li. For disabled
elements (prev, next=, the span tag is used and also added the
active class for the current page.