Update bake code generation to CakePHP v.2.3.1
Updated the custom VFXfan bake code generation to CakePHP v2.3.1.

Views were mostly cosmetic changes, like eliminating the subpackage
doc comment. The index views also changed the pagination block to
use the new class and output that works with the default Bootstrap
pagination styles.

Classes also had minor cosmetic changes, like eliminating the
subpackage doc comment. The model class fixed the $useDbConfig
inequality test.

The controller actions file received the most changes. It now uses
a more compact way to check if a record exists, now that the
exists() model method accepts an id parameter: in the view and edit
methods. When searching for a record, we separate the
options into an array that we pass to a find('first') call: in the
view and edit methods. In the delete method we use the new
$this->request->onlyAllow() call to check that only post and delete
requests are permitted, if not it throws a MethodNotAllowedException
and we can eliminate that additional check.