Update base systen to CakePHP v.2.3.1
Updated the VFXfan base system to CakePHP v2.3.1.

This was based on the updates to the bake code generation.

Methods now use a more compact way to check if a record exists,
now that the exists() model method accepts an id parameter: mainly
in the view and edit methods. When searching for a record, we
separate the options into an array that we pass to a find('first')
call, mainly in the view and edit methods. In the delete method we
use the new $this->request->onlyAllow() call to check that only
post and delete requests are permitted, if not it throws a
MethodNotAllowedException and we can eliminate that additional
check. When uploading or deleting files we also check that the
record exists.

In the models we also check that a record exists in the afterSave
and beforeDelete callbacks.