Fix display of Albums images
Fix the display of thumbnails and linked Albums images. Classes were
changed to correctly display thumbnails. Thumbnails use the
.img-thumbnail class. For the border we already have the CSS in the
skin customizations. The grids now have the .text-center class.

Eliminated the Slimbox2 references and plugin and used the Magnific
Popup jQuery plugin instead. Added the calls in the project and admin
Javascript files and added the necessary classes to display the
lightboxes. For the gallery an eclosing div is necessary.