Fix Paginator settings
Fixed the Paginator settings. In the Pages Controller the class
incorrectly set the limit, which was overriden in the view method.

For the other controllers the options were set inside the pagination
call that also set the view variable. We separated the options to
the settings call for better control and code cleanup.