Fix admin_delete method with unset record id
Fixed the admin_delete method which tried to delete a record with the
empty delete() call, but the id was not set before. We use the bake
action as an example, so now we set the id first with a call to:

$this->Model->id = $id;

After that we can both check the existance of a record and delete it
with empty calls exists() and delete().