Fix menu items call for performance and validity
Couple of minor changes to make the navigation menu better. First use
array parameters in the requestAction() call in the defaul layout so
we can skip the Router, making it a bit more efficient. Second, verify
in the controller that the call actually originated from a
requestAction(), so it isn't accesible via an URL, throw a Forbidden
exception if it wasn't a requested action.