Make media files contained in their own folder
Changed several systems so that media files in the webroot, are
contained in their own directory. This makes it easier to manage the
media files, like images, videos and PDF files, by isolating them
from the year folders. The new containing folder has a name formed by
the zero justified id of the record, like 0000000001.

In the model we modify the afterSave and beforeDelete callbacks to
create and delete the new folders. The controllers were changed in
their admin add and edit methods to upload the images to the new
folders. The view were changed to display the media in their new
locations.

For the Albums we also create a new preview folder which is used for
the album thumbnail image.