Compile skin CSS to build folder
Changed the Gruntfile so that the CSS skin is compiled, linted and
minified to the build folder, as opposed to the tmp folder. It's also
copied to the project webroot. This is done so that we can use the
Bootstrap CDN and just include the skin in our configuration and
layouts.