Add skin CSS and Bootstrap CDN to layouts
Since Bootstrap is now available via MaxCDN, we use the Bootstrap CDN
links in the layouts. The links are actually specified in the project
configuration files and added to the layouts by checking for their
existence.

Since we're using the Bootstrap CDN, we no longer need to include it
in our project or compile a version that incorporates our skin, so
we just include the in the layouts, which is now compiled and copied
via Grunt.