Use customized compiled Bootstrap files
Use customized compiled versions of the Bootstrap files. This will
simplify the skin Less files and the frontend CSS development. Since
we are customizing Bootystrap we won't be able to use the CDN
version. We need to compile the skin and concatenate it to the
Bootstrap CSS to create the project CSS.

The basic variables to customize are:

@headings-font-family: 'Open Sans', Helvetica, Arial, sans-serif;
@headings-color: @brand-primary;

We could also customize @link-color and @navbar-default-brand-color.

The link color is still defined in the skin-variables file since we
use it for the img-thumbail styling.

New colors can be defined by using the Paletton website:

http://paletton.com/

Other tools to generate collor palettes are:

- https://kuler.adobe.com/en/create/color-wheel/

- http://colllor.com/