Update to CakePHP v2.5.3
Updated the system to CakePHP v2.5.3.  For the view files and app
index, the fixes were only fomatting issues.

On the other hand most config files do have some changes.

There is an error on the db_acl.sql file which uses an incorrect name
for the right index of the AROs and ACOs tables. The index uses the
name rhgt instead of rght, which will choke the database on manual
update. Had to correct it manually and created an issue in GitHub
with number #4052.