Add functionality to search preview image sizes in content
Added the functionality to seach preview images in a specific order.

First in the model we created the class property $orderPreviewImages
where we list in order the image sizes prefix. Then in the
afterFind() callback we search for images with these extensions and
append the results into the preview_images key of the result set.

Also modified the Models to also include other images in their preview
folders which are not part of the new naming convention for preview
image sizes.

We use a regular expresion which uses a negative lookaround to make
sure any of the size specifiers don't precede the file extension.
We then add the list of files to the end of the preview images
subarray.

This was based in the following StackOverflow posts:

http://stackoverflow.com/questions/12437757/is-there-a-such-thing-as-negated-alternation-or-negated-string-classes-a-l
http://stackoverflow.com/questions/406230/regular-expression-to-match-string-not-containing-a-word