Add method to get preview image
Added a method to the FormatImage Helper to get the name of a preview
image. The method, getPreviewImage(), takes two parameters, an array
with the names of all preview images, and a an options array. If the
size is specified it tries to find an image. If no size is specified
the method iterates over the array and returns the first image it
finds. If not, it returns an empty string.