Add style for half page sections and content preview
Added CSS classes to display content lists with preview images. The
preview image will be in a 4 column section and the content in an 8
column section.

This will be used for the index views for Posts and Events.

Added the style half-page-section to the skin, which is equivalent to
col-md-6 so different section can occupy half the page. We delete the
id elements which are no longer necessary for styling.