Add Font Awesome support
Added support for Font Awesome, bu adding a configuration variable in
the project.php file and adding a check in the default layout.

We also reorganized the order in which CSS files are loaded. We first
load jQuery UI file, then Fon Awesome, then the Bootstrap files and
last any other CSS files.