Refactor the file delete admin function for Pages
Changed the functionality of the admin_deleteFile method in the Pages
controller to eliminate some duplicated code and make it a bit more
flexible.

Set the default location to 'img', no need to check for a string if
we can manage with the file location. Also added a redirect_action
parameter. This is because we always redirected to the admin_view
action, even if we had deleted a file from the admin_edit action.

We combined the two if statements to check which file to delete
(image or document) into one and set the appropiate flash message.
Finally we redirect based on the redirect_action parameter.

In the admin_edit view, and we add the appropiate parameters to
redirect and delete the correct type of file.