Modify delete file functionality to use named parameters
Modified the functionality for deleteing files so we can delete both
preview images and files.

In the views we use named parameters, instead of positional
parameters, so it's easier to pass information to the controller and
model. We pass the file name, the redirect action and the base folder
of the file with respect to the webroot.

In the controller we don't need the filename parameter anymore and
we use the named parameters to pass the information to the model and
also set the redirect action.