Change Form->value calls to use CakeRequest
Changed the calls of Form->value() to use the CakeRequest object,
available as $this->request in views. Form->value() was used to
get the id of the record, present in most edit and a few add views.
The value() call is actually inherited from the base Helper. Since
the id is available in the request object anyway, we use that. This
call was mostly used in buttons and dropdows to delete files and
records.