Use the data method instead of the property of CakeRequest
Changed the use of the data property of the Cakeresponse object, to
the data method instead. That is, instead of $this->request['data'],
we use $this->request->data() instead. This is because it's more
error free, because it can return null on non existent keys, and it
also uses dot notation.