Fix adding extra page sections
Fixed the functionality for adding extra Page Sections. The problem
was that since all extra sections had the same names and ids, only
the last one would be saved.

First in the admin_add Pages view we add the class page-section-title
to the label for the section title input. We use his dummy CSS class
to count how many page sections we have.

Then in the admin.js file we add this class to all the new Page
Sections that are appended.