Fix styling and output of checkbox
Fixed the styling of the checkboxes in by changing how it is
generated so that it better matches the way Bootstrap wants
checkboxes to be coded.

We eliminate the automatic label by setting the property false and
then add it back using the before and after properties of the input
call.