Upgrade to Bootstrap v3.3.0
Upgraded frontend to Bootstrap v3.3.0 and generated new CSS files
and also copied the corresponding Javascript, Less and font files.

We use a customized Bootstrap file to generate the compiled files.
The custom variables are the same as those in the v3.2.0:

@link-color: #d5000b;
@heading-font-family: 'Electrolize', Helvetica, Arial, sans-serif;
@heading-color: #0b3891;
@navbar-default-brand-color: #0b3891;

and using the @link-color variable in the skin as well.

The configuration has been saved to:

https://gist.github.com/f23fd26dcf0e405817c6

and can be revisited at:

http://getbootstrap.com/customize/?id=f23fd26dcf0e405817c6

for further customization.