Add first version of AppModel test case
Added the first version of the AppModel test case to check the
available methods, mainly dealing with cleaning input.

Created tests for alphaNumericDashUnderscoreSpaceColon() and
alphaNumericDashUnderscore(). Also created a test for the purify()
which uses HTMLPurifier to clean the input. Created the assertions
using the HTMLPurifier smokestack, which is based on the OWASP XSS
Filter Evasion Cheat Sheet:

http://htmlpurifier.org/live/smoketests/xssAttacks.php
https://www.owasp.org/index.php/XSS_Filter_Evasion_Cheat_Sheet

A few tests were not implemented, those that use null characters and
those that blacklist a domain (Google in the case of the examples).

Added the AllModelTest test suite to test all models. Added two test
other suites: one to run all present tests and one to test all
helpers.