Simplify code to add new Page Sections on new Page
Simplified the code to create new extra Page sections when adding a
new Page. First we restructured the admin_add Page view. We
separate the first page section in its own fieldset tag to help
in duplicating it. This fieldset has an id of page-section-0,
to find it and clone it, and a class of page-section, to count the
number of Page Sections. Also instead of having a link in the
actions dropdown button, we use a button to add new Page Sections
in the div where they will be appended.

In the admin jQuery file we clone the first Page Section fieldset
and then use the find() call to locate the descendat form fields
and modify the corresponding attributes like for (for labels), name
and id. The new fieldset is then appended to the div for extra Page
Sections. We also use variables to help construct the new
attributes.

We also use a strict comparison to count the number of extra Page
Sections.