Refactor Upload Component
We refactored the Upload component to reduce redundant code and
simplify its operation. Instead of two upload methods (one for images
and one for files), we only have one uploadFile() method.

The create thumbnail functionality has been in it's own method so we
can call it if the option is set to truth. The private variables for
permissible MIME types are gone and now we provide a default option
(for images), in the options of the checkTypes() method.

Modified the Upload component so that we have the option to upload
an image, create the thumbail and store it, but not uploading the
original full size image.

We create two more options: thumbs_folder to determine if we should
use the thumbnails folder in the specified pass, and only_thumbnail
to specifiy if we only want to create/store the thumbnail image.
This can be useful when creating preview images for posts.