First version of SimpleRecaptcha plugin
Commiting first version of the SimpleRecaptcha Plugin. This is
partially based on the CakeDC plugin, particulalrly the Component
initialize method. We use the reCAPTCHA API which was modified on
Novermber 19, 2014.

In the Component we have two methods. In the initialize method we make
sure to add the Helper to the calling controller and verify that the
private key is set. We also add the reCAPTCHA textarea input to the
list of unlocked fileds, since it won't be part of the original form
wherever it's used. The verify method get the reposnse string from the
widget and makes and HTTPS request to the reCAPTCHA API to verify it.
The response is a JSON object which comes as a string in the body
attribute of the response object, so we need to decode it first. We
retrun this so that the success status can be verified in the calling
controller.

For the Helper we just display a widget by using a div with the public
key and the necessary class.