Add SimpleRecaptcha to Contact Us form
Added the new SimpleRecaptcha plugin to the Contact Forms system. We
just add the Google Javascript API to the index view where the form
is and enable the plugin in the project bootstrap. In the controller
we just verify that the response string hass a success value of true
before we save any data