Add method to upload several files at a time
Added the uploadFiles() method to the upload several files at the
same time when using the multiple upload files capabilities of HTML5.
This methos just loops over the data array and call the uploadFile()
method of the component passing the same parameters.

To use in forms, we must unlock the file upload fileds in the
beforeFilter() callback (without the dot to indicate an array). In
the file upload field, using the input method of the Form helper, we
add the multiple attribute, and the field name must end with a dot
so that the name includes and extra pair of square brackets to supprt
uploading several files. We must also set a labe for the file input
field.