Add multiple file upload capability to Pages
Added the HTML5 multiple file upload capability to the Pages system.
In the controller we make sure to unlock the documents and images
file upload fields, using the name without the ending dot, for the
admin_edit method. In the admin_edit action we change the call from
uploadFile() to uploadFiles() with the exact same parameters.

In the view we add the multiple attribute to the file upload inputs.
For the name we add a dot, so that it creates a properly formated
field. For example the field name changes from 'File.image' to
'File.image' so the field name becomes 'data[File][image][]' (notice
the empty square brackets at the end). We also add a label explicitly,
otherwise the text is empty.