Modify convertFilenameToId method to handle more extensions
Modified the convertFilenameToId() method of the Upload component to
handle more cases, in particular when there are more than one
extension, for example, for resposive images. We now use the pathinfo
function to get the last extension, and then use the conversion on the
rest of the file name. Now, a file like 'star-wars.md.jpg' will be
converted to something like '0000000001.md.jpg'.

We also updated the test case to reflect the improved functionality.