First version of Resize Image component
Created first version of the ResizeImage component, to take out that
functionality out of the Upload component. It is based on the
following TutsPlus tutorial:

http://code.tutsplus.com/tutorials/image-resizing-made-easy-with-php--net-10362

with a few modifications. One important difference is that the crop
option has been eliminated.

One differnce is that we use a regular constructor to initialize an
instance, instead of a CakePHP component constructor, which takes a
ComponentCollection instance and a settings array. Since we are only
using this component insode the Upload component, and not in
controllers, it's not a huge issue. In the Upload component we only
need to create an instance and pass the file and MIME type.

The class creates an image object. To resize the image we get an
optimal size based on the desired dimensions and keeping the image
proportional. There are two parts to the resize method. The first is
getting the optimal width and height for our new image by creating
some custom methods, and of course passing in a resize option. The
width and height are returned as an array and set to their
respective variables. We use imagecreatetruecolor and
imagecopyresampled to create the resized image and save it to a
class variable. We finally save the image out based on MIME type.

Also added a method to create a series of different sized images for
responsive purposes, called responsiveImages. It takes an array of
sizes to create all the sizes. The key is used as an additional
extension to the file name, and the value is an int value specifying
the width. Responsive images are resized only based on width.

Some other modifications included renaming some variables to be more
meaningful.