Use Resize Image component in Upload component
Simplified the Upload component by taking out the previously image
resizing functionality and using the new ResizeImage component. There
are two methods now, one to create a single thumnail image and another
to create a series of responsive images based on width.

There are a few new options, including one to specify if we want to
create the responsive images and a sizes array to specify the widths
and extensions of the responsive images.