Add support to upload multiple files on Page creation
Added support to upload both several images and files when adding a
new Page. We modified the form to support file uploads and added the
fields to upload multiple files and images.

In the controller we modifiy the beforeFilter() callback so that we
don't validate Page Sections on the admin_add method as well. We add
the images and documents to the list of unlocked fields.

In the admin_add method we call the uploadFiles method of the Upload
component to upload files, similar to the admin_edit method. The main
difference is we use $this->Page->id instead, just available after
saving the Page, to get the proper id.