Change variable name of unlocked form fields
Changed the varibale name used to unlock form fields in the Pages
controller from $unlocked_add_fields to $unlocked_fields.

The previous name was used to indicate that it was used for extra
(dynamically added) for fields, that should not be checked on submit
by the Security component. Now we are also unlocking fields used to
submit multiple files for upload.