Eliminate constructor from Resize Image component
Eliminated the cosntructor for the ImageResize component. This is
because instancing the component the component directly still calls
the callback, producing a silent error, though it still resized the
images (brain fart).

The constructor just opened an image and determined the size of the
original image, so we moved that to the openImage() method, and also
make it public.