Create drag and drop AJAX file upload area for Page files
Started development of an AJAX file upload drag and drop area for
Pages files. This is another alternative to choosing several files
in the form to submit.

First we create a div in the admin_edit view, below the form so it's
clear it separate from the form submission. We add a class of
dropzone and an id of dropzone-files we can target.

We style the div so that it has a gray background, with a dashed
outline and centered text. It will change to a darker gray when we
over files.

This is based on the Udemy course 'Create a Drag & Drop File Uploader
For Websites' from Stone River eLearning:

https://www.udemy.com/create-a-udemy-style-drag-drop-file-uploader-for-websites

which might actually be based on a PHPAcademy video.