Add data attributes for AJAX file uploads
Added data attributes to the files drag and drop area, which we will
need to upload the files to the correct directory. We need the base
URL (data-base), Controller (data-controller)m record id (data-id)
and file type (data-file-type). We use the file type to determine if
we are uploading images or documents. For the record id we use the
first element od the passed parameters in the admin_edit forms, since
this is the first (and only) parameter we use.

In the admin Javascript file we check in the handler for the drop
event if we have a data transfer property and if so, we check if the
files list length is not zero. We then pass the data attributes to
the uploadFiles function.

In the meantime we create an uploadFiles function and we console log
the files to check it's working.