Add AJAX uploads for Pages files
Added the AJAX upload functionality for Pages files. In the admin
Javascript we create an uploadFiles function that takes five
parameters: the files, base URL, Controller, record id and upload
type. We decided to switch the name of the data attribute in the
form from data-file-type to data-upload-type because if we implement
it in other parts of the app, we might have to distinguish between
regular images and image previews. We reflect this change in the
drop event handler as well.

In the uploadFiles function we create a FormData object to submit
the files, that way we can use the regular Upload component. We loop
over the files and append them to the FormData object. We then use
the jQuery ajax function to submit the FormData object using a POST
type of request. We set the cache, contentType and processData
options to false so that jQuery doesn't convert or process the data.
On success we reload the page. While we could probably send back more
data, say the list of files uploaded or a progress indicator, we just
reload the page so we can see the updated list of files.

In the controller we add a new admin_ajaxUploadFiles() method to
actually handle the uploads. We set the render to false so it doesn't
use a view file we check to see if the id is set. We then call the
Upload component using a switch statement that depends on the upload
type parameter to select exactly where the file should go. Since the
files come as if they were submitted by a form, CakePHP's Security
component would check for form tampering, so we need to unlock the
method in the beforeFilter callback. Also we must run the ACL ACO
update. After the upload we sennd a 200 status code response back
to signify success (even though we don't check for errors) to reload
the page.