Add AJAX upload functionality for Pages images
Added the functionality for AJAX uploading for Pages images via a
drag and drop area. We just duplicated the functionality we did
for the files uploads and adjusted it. Mainly in the view we change
the area to have a different id and the data-upload-type changes to
images.

In the admin Javscript we add the dropzone-images id for the event
handlers. In the Pages controller we add a case to the switch
statement of the admin_ajaxUploadFiles() method to handle the images
uploads.