Add options parameter for AJAX drag and drop file uploads
Added the ability to use an options array (hash) parameter when doing
AJAX drag and drop file uploads. In the drop event handler we just
the value of a data attribute called data-options, and pass it to the
uploadFiles() function. In case the data attribute is missing the
value will be null. This data attribute should contain a JSON encoded
PHP array, for example:

htmlspecialchars(json_encode(array('year' => '1977')))

and the data attribute value should be enclosed in single quotes,
since the JSON values must be enclosed in double quotes.

In the uploadFiles() function (where we added an options parameter),
we convert the options to a string value using the JSON.stringify()
function, and append this to the formData variable.

The options will be available in the $this->request->data['options']
property, which needs to be JSON decoded like:

json_decode($this->request->data['options'])->{'year'}

which we can now use.

Information on JSON.stringify:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/JSON/stringify

Some useful Stackoverflow posts and Gists that helped figure this
out:

https://gist.github.com/charliepark/4266921
http://stackoverflow.com/questions/8832528/escaping-encoding-single-quotes-in-json-encoded-html5-data-attributes
http://stackoverflow.com/questions/19976627/posting-json-with-jquery-ajax-to-php
http://stackoverflow.com/questions/6974684/how-to-send-formdata-objects-with-ajax-requests-in-jquery
http://stackoverflow.com/questions/5046930/jquery-send-string-as-post-parameters